changed xml to be complient with java; added new xml file.
Java doesn't like it when an XML file refers to a DTD without the dtd
extension, so I added it. I also created an xml file,
`currentLayout.xml', which at the moment is just the same as
averyAddressLabel.xml. The idea is that XML files describing layouts are
kept with the name of the layout, but the one you want to use at the
moment is copied into a new file with the name `currentLayout.xml'